app/requirements.txt
streamlit
openai
psycopg2-binary
sqlalchemy
bcrypt
python-dotenv
